+++ apps/design-system/content/docs/fragments/multi-select-old.mdx
@@ -5,7 +5,7 @@ component: true
 fragment: true
 ---
 
-<ComponentPreview name="multi-select-demo" description="" peekCode wide />
+<ComponentPreview name="multi-select-old-demo" description="" peekCode wide />
 
 ## Usage
 
@@ -17,7 +17,7 @@ import {
   MultiSelectorItem,
   MultiSelectorList,
   MultiSelectorTrigger,
-} from 'ui-patterns/multi-select'
+} from 'ui-patterns/multi-select-old'
 ```
 
 ```tsx
@@ -39,4 +39,4 @@ import {
 
 ### Form
 
-<ComponentPreview name="multi-select-form" />
+<ComponentPreview name="multi-select-old-form" />
