+++ apps/design-system/content/docs/fragments/multi-select.mdx
@@ -5,7 +5,7 @@ component: true
 fragment: true
 ---
 
-<ComponentPreview name="multi-select-2-demo" description="" peekCode wide />
+<ComponentPreview name="multi-select-demo" description="" peekCode wide />
 
 ## Usage
 
@@ -15,7 +15,7 @@ import {
   MultiSelectorTrigger,
   MultiSelectorContent,
   MultiSelectorItem,
-} from 'ui-patterns/multi-select-2'
+} from 'ui-patterns/multi-select'
 ```
 
 ```tsx
@@ -39,20 +39,20 @@ import {
 
 ### Badge Combo Box
 
-<ComponentPreview name="multi-select-2-combobox" />
+<ComponentPreview name="multi-select-combobox" />
 
 ### Badge Limit
 
-<ComponentPreview name="multi-select-2-badge-limit" />
+<ComponentPreview name="multi-select-badge-limit" />
 
 ### Badge Limit="auto"
 
-<ComponentPreview name="multi-select-2-badge-limit-auto" />
+<ComponentPreview name="multi-select-badge-limit-auto" />
 
 ### Badge Limit="wrap"
 
-<ComponentPreview name="multi-select-2-badge-limit-wrap" />
+<ComponentPreview name="multi-select-badge-limit-wrap" />
 
 ### Deletable Badge
 
-<ComponentPreview name="multi-select-2-deletable-badge" />
+<ComponentPreview name="multi-select-deletable-badge" />
