+++ apps/design-system/registry/examples.ts
@@ -1031,44 +1031,44 @@ export const examples: Registry = [
     files: ['example/info-tooltip-demo.tsx'],
   },
   {
+    name: 'multi-select-old-demo',
+    type: 'components:example',
+    files: ['example/multi-select-old-demo.tsx'],
+  },
+  {
+    name: 'multi-select-old-form',
+    type: 'components:example',
+    files: ['example/multi-select-old-form.tsx'],
+  },
+  {
     name: 'multi-select-demo',
     type: 'components:example',
     files: ['example/multi-select-demo.tsx'],
   },
   {
-    name: 'multi-select-form',
+    name: 'multi-select-badge-limit-auto',
     type: 'components:example',
-    files: ['example/multi-select-form.tsx'],
+    files: ['example/multi-select-badge-limit-auto.tsx'],
   },
   {
-    name: 'multi-select-2-demo',
+    name: 'multi-select-badge-limit-wrap',
     type: 'components:example',
-    files: ['example/multi-select-2-demo.tsx'],
+    files: ['example/multi-select-badge-limit-wrap.tsx'],
   },
   {
-    name: 'multi-select-2-badge-limit-auto',
+    name: 'multi-select-badge-limit',
     type: 'components:example',
-    files: ['example/multi-select-2-badge-limit-auto.tsx'],
+    files: ['example/multi-select-badge-limit.tsx'],
   },
   {
-    name: 'multi-select-2-badge-limit-wrap',
+    name: 'multi-select-combobox',
     type: 'components:example',
-    files: ['example/multi-select-2-badge-limit-wrap.tsx'],
+    files: ['example/multi-select-combobox.tsx'],
   },
   {
-    name: 'multi-select-2-badge-limit',
+    name: 'multi-select-deletable-badge',
     type: 'components:example',
-    files: ['example/multi-select-2-badge-limit.tsx'],
-  },
-  {
-    name: 'multi-select-2-combobox',
-    type: 'components:example',
-    files: ['example/multi-select-2-combobox.tsx'],
-  },
-  {
-    name: 'multi-select-2-deletable-badge',
-    type: 'components:example',
-    files: ['example/multi-select-2-deletable-badge.tsx'],
+    files: ['example/multi-select-deletable-badge.tsx'],
   },
   {
     name: 'tree-view-demo',
