--- packages/ui-patterns/multi-select-2/index.tsx
@@ -1,1 +1,0 @@
-export * from './multi-select-2'
